- Minor refactoring - snakecase - Remove unused methods
