+++ core/src/stream/end_line.rs
