+++ core/src/stream/mod.rs
@@ -16,7 +16,7 @@
 
 pub mod array_stream;
 pub mod char_stream;
-pub mod endline;
+pub mod end_line;
 pub mod iterator_stream;
 pub mod parser_stream;
 pub mod position;
